Move terrain textures loading to own file
